pyproject.toml
[project]
name = "agent-assitant"
version = "0.1.0"
description = "Layered FastAPI server with OpenAI-backed agent chat"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=1.70.0",
    "pydantic-settings>=2.8.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.44",
    "psycopg[binary]>=3.2.12",
    "PyJWT>=2.10.1",
    "bcrypt>=5.0.0",
    "uvicorn>=0.42.0",
    "pinecone>=8.1.0",
    "qdrant-client>=1.12.0",
    "minio>=7.2.18",
    "pika>=1.3.2",
    "python-multipart>=0.0.20",
    "ipykernel>=7.2.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.12.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "sympy>=1.13.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=8.3.5",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
pythonpath = ["."]
